Implements deleted second level comments and refactor util to only accept id.

+++ components/Comments/CommentContent.tsx
@@ -51,12 +51,24 @@ const CommentContent: React.FC<CommentContentProps> = ({
 	const deleteHandler: MouseEventHandler<HTMLButtonElement> = async () => {
 		if (confirm("Are you sure you want to delete this comment?")) {
 			try {
-				await fetch(`/api/comments/${currComment._id}`, {
-					method: "DELETE",
-					headers: {
-						"Content-Type": "application/json",
-					},
-				});
+				if (isSecondLevelComment) {
+					if (originalComment && originalComment.replies) {
+						const deletedComment = {
+							...originalComment,
+							replies: originalComment.replies.filter(
+								(reply) => reply._id !== currComment._id
+							),
+						};
+						postUpdatedResourceToDb(deletedComment, originalComment._id);
+					}
+				} else {
+					await fetch(`/api/comments/${currComment._id}`, {
+						method: "DELETE",
+						headers: {
+							"Content-Type": "application/json",
+						},
+					});
+				}
 
 				await scrollToSamePosition();
 			} catch (error) {
@@ -165,10 +177,10 @@ const CommentContent: React.FC<CommentContentProps> = ({
 
 						if (isSecondLevelComment) {
 							if (originalComment) {
-								await postUpdatedResourceToDb(commentBody, originalComment);
+								await postUpdatedResourceToDb(commentBody, originalComment._id);
 							}
 						} else {
-							await postUpdatedResourceToDb(commentBody, currComment);
+							await postUpdatedResourceToDb(commentBody, currComment._id);
 						}
 					} catch (error) {
 						console.error(error);
